+++ Helpers/ZendeskGet.cs
@@ -56,7 +56,7 @@ namespace Helpers {
                 throw ZendeskError.FromErrors(response.Errors);
         }
 
-        public async static Task<ZendeskSell.Orders.OrderResponse> GetOrder(ZendeskSell.Orders.IOrderActions orderActions, int dealID) {
+        public async static Task<ZendeskSell.Orders.OrderResponse> GetOrder(ZendeskSell.Orders.IOrderActions orderActions, long dealID) {
             var orders = await GetAll((pn, pc) => orderActions.GetAsync(pn, pc, dealID));
             if (orders.Count() == 0)
                 return Handle(await orderActions.CreateAsync(new ZendeskSell.Orders.OrderRequest() {
